EPMRPP-112183-remove-github || Refactor authentication integration and update package structure
